Review and refactor
