+++ core/src/test/kotlin/com/strumenta/kolasu/mapping/ParseTreeToASTTransformerTest.kt
@@ -3,21 +3,52 @@ package com.strumenta.kolasu.mapping
 import com.strumenta.kolasu.model.*
 import com.strumenta.kolasu.parsing.withParseTreeNode
 import com.strumenta.kolasu.testing.assertASTsAreEqual
-import com.strumenta.kolasu.transformation.ASTTransformer
-import com.strumenta.kolasu.transformation.GenericNode
-import com.strumenta.simplelang.SimpleLangLexer
-import com.strumenta.simplelang.SimpleLangParser
-import org.antlr.v4.runtime.CharStreams
-import org.antlr.v4.runtime.CommonTokenStream
+import com.strumenta.kolasu.transformation.*
+import com.strumenta.kolasu.traversing.walk
+import com.strumenta.kolasu.validation.Issue
+import com.strumenta.simplelang.*
+import org.antlr.v4.runtime.*
+import org.antlr.v4.runtime.atn.ATNConfigSet
+import org.antlr.v4.runtime.dfa.DFA
 import org.junit.Test
+import java.util.*
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
+import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
 data class CU(val specifiedPosition: Position? = null, var statements: List<Node> = listOf()) : Node(specifiedPosition)
 data class DisplayIntStatement(val specifiedPosition: Position? = null, val value: Int) : Node(specifiedPosition)
 data class SetStatement(val specifiedPosition: Position? = null, var variable: String = "", val value: Int = 0) :
     Node(specifiedPosition)
+
+data class EModule(override val name: String, val entities: MutableList<EEntity>) : Node(), Named
+data class EEntity(override val name: String, val features: MutableList<EFeature>) : Node(), Named
+
+data class EFeature(override val name: String, val type: EType) : Node(), Named
+
+sealed class EType : Node()
+class EStringType : EType()
+class EBooleanType : EType()
+data class EEntityRefType(val entity: ReferenceByName<EEntity>) : EType()
+
+data class SScript(val statements: MutableList<SStatement>) : Node()
+sealed class SStatement : Node()
+data class SCreateStatement(val entity: ReferenceByName<EEntity>, val name: String? = null) : SStatement()
+data class SSetStatement(val feature: ReferenceByName<EFeature>, val instance: SExpression, val value: SExpression) :
+    SStatement()
+data class SPrintStatement(val message: SExpression) : SStatement()
+
+sealed class SExpression : Node()
+data class SStringLiteral(val value: String) : SExpression()
+data class SIntegerLiteral(val value: Int) : SExpression()
+data class SDivision(val left: SExpression, val right: SExpression) : SExpression()
+data class SSubtraction(val left: SExpression, val right: SExpression) : SExpression()
+data class SMultiplication(val left: SExpression, val right: SExpression) : SExpression()
+data class SSum(val left: SExpression, val right: SExpression) : SExpression()
+data class SConcat(val left: SExpression, val right: SExpression) : SExpression()
+data class SFeatureAccess(val feature: ReferenceByName<EFeature>, val container: SExpression) : SExpression()
+data class SInstanceById(val entity: ReferenceByName<EEntity>, val index: SExpression) : SExpression()
 
 class ParseTreeToASTTransformerTest {
 
@@ -120,4 +151,289 @@ class ParseTreeToASTTransformerTest {
             SetStatement(variable = ctx.ID().text, value = ctx.expression().INT_LIT().text.toInt())
         }
     }
+    class MyErrorListener : ANTLRErrorListener {
+        override fun syntaxError(
+            p0: Recognizer<*, *>?,
+            p1: Any?,
+            line: Int,
+            column: Int,
+            message: String?,
+            p5: RecognitionException?
+        ) {
+            throw RuntimeException("L$line:$column: $message")
+        }
+
+        override fun reportAmbiguity(
+            p0: Parser?,
+            p1: DFA?,
+            p2: Int,
+            p3: Int,
+            p4: Boolean,
+            p5: BitSet?,
+            p6: ATNConfigSet?
+        ) {
+            // nothing to do
+        }
+
+        override fun reportAttemptingFullContext(
+            p0: Parser?,
+            p1: DFA?,
+            p2: Int,
+            p3: Int,
+            p4: BitSet?,
+            p5: ATNConfigSet?
+        ) {
+            // nothing to do
+        }
+
+        override fun reportContextSensitivity(p0: Parser?, p1: DFA?, p2: Int, p3: Int, p4: Int, p5: ATNConfigSet?) {
+            // nothing to do
+        }
+    }
+
+    private fun parseEntities(code: String): AntlrEntityParser.ModuleContext {
+        val errorListener = MyErrorListener()
+
+        val lexer = AntlrEntityLexer(CharStreams.fromString(code))
+        lexer.removeErrorListeners()
+        lexer.addErrorListener(errorListener)
+        val parser = AntlrEntityParser(CommonTokenStream(lexer))
+        parser.removeErrorListeners()
+        parser.addErrorListener(errorListener)
+        return parser.module()
+    }
+
+    private fun parseScript(code: String): AntlrScriptParser.ScriptContext {
+        val errorListener = MyErrorListener()
+
+        val lexer = AntlrScriptLexer(CharStreams.fromString(code))
+        lexer.removeErrorListeners()
+        lexer.addErrorListener(errorListener)
+        val parser = AntlrScriptParser(CommonTokenStream(lexer))
+        parser.removeErrorListeners()
+        parser.addErrorListener(errorListener)
+        return parser.script()
+    }
+
+    @Test
+    fun testSimpleEntitiesTransformer() {
+        val transformer = ParseTreeToASTTransformer(allowGenericNode = false)
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.ModuleContext, EModule>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.EntityContext, EEntity>()
+        val expectedAST = EModule(
+            "M",
+            mutableListOf(
+                EEntity("FOO", mutableListOf()),
+                EEntity("BAR", mutableListOf())
+            )
+        )
+        val actualAST = transformer.transform(
+            parseEntities(
+                """
+            module M {
+               entity FOO { }
+               entity BAR { }
+            }
+                """.trimIndent()
+            )
+        )!!
+        assertASTsAreEqual(expectedAST, actualAST)
+    }
+
+    @Test
+    fun testEntitiesWithFeaturesTransformer() {
+        val transformer = ParseTreeToASTTransformer(allowGenericNode = false)
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.ModuleContext, EModule>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.EntityContext, EEntity>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.FeatureContext, EFeature>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.String_typeContext, EStringType>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.Boolean_typeContext, EBooleanType>()
+        transformer.registerTrivialPTtoASTConversion<AntlrEntityParser.Entity_typeContext, EEntityRefType>(
+            "target" to "entity"
+        )
+        val expectedAST = EModule(
+            "M",
+            mutableListOf(
+                EEntity(
+                    "FOO",
+                    mutableListOf(
+                        EFeature("A", EStringType()),
+                        EFeature("B", EBooleanType()),
+                    )
+                ),
+                EEntity(
+                    "BAR",
+                    mutableListOf(
+                        EFeature("C", EEntityRefType(ReferenceByName("FOO")))
+                    )
+                ),
+            )
+        )
+        val actualAST = transformer.transform(
+            parseEntities(
+                """
+            module M {
+               entity FOO {
+                A: string;
+                B: boolean;
+                 
+               }
+               entity BAR {
+                 C: FOO;
+               }
+            }
+                """.trimIndent()
+            )
+        )!!
+        assertASTsAreEqual(expectedAST, actualAST)
+    }
+
+    @Test
+    fun testScriptTransformer() {
+        val transformer = ParseTreeToASTTransformer(allowGenericNode = false)
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.ScriptContext, SScript>()
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Create_statementContext, SCreateStatement>(
+            AntlrScriptParser.Create_statementContext::var_name to SCreateStatement::name
+        )
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Set_statementContext, SSetStatement>()
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Entity_by_id_expressionContext, SInstanceById>(
+            AntlrScriptParser.Entity_by_id_expressionContext::id to SInstanceById::index
+        )
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Int_literal_expressionContext, SIntegerLiteral>(
+            AntlrScriptParser.Int_literal_expressionContext::INT_VALUE to SIntegerLiteral::value
+        )
+        transformer.registerNodeFactory(AntlrScriptParser.String_literal_expressionContext::class) { pt, t ->
+            SStringLiteral(pt.text.removePrefix("'").removeSuffix("'"))
+        }
+        transformer.registerNodeFactory(AntlrScriptParser.Div_mult_expressionContext::class) { pt, t ->
+            when (pt.op.text) {
+                "/" -> {
+                    TrivialFactoryOfParseTreeToASTNodeFactory.trivialFactory<
+                        AntlrScriptParser.Div_mult_expressionContext, SDivision>()(pt, t)
+                }
+                "*" -> {
+                    TrivialFactoryOfParseTreeToASTNodeFactory.trivialFactory<
+                        AntlrScriptParser.Div_mult_expressionContext, SMultiplication>()(pt, t)
+                }
+                else -> TODO()
+            }
+        }
+        transformer.registerNodeFactory(AntlrScriptParser.Sum_sub_expressionContext::class) { pt, t ->
+            when (pt.op.text) {
+                "+" -> {
+                    TrivialFactoryOfParseTreeToASTNodeFactory.trivialFactory<
+                        AntlrScriptParser.Sum_sub_expressionContext, SSum>()(pt, t)
+                }
+                "-" -> {
+                    TrivialFactoryOfParseTreeToASTNodeFactory.trivialFactory<
+                        AntlrScriptParser.Sum_sub_expressionContext, SSubtraction>()(pt, t)
+                }
+                else -> TODO()
+            }
+        }
+        transformer.unwrap<AntlrScriptParser.Parens_expressionContext, SExpression>(AntlrScriptParser.Parens_expressionContext::expression)
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Print_statementContext, SPrintStatement>()
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Concat_expressionContext, SConcat>()
+        transformer.registerTrivialPTtoASTConversion<AntlrScriptParser.Feature_access_expressionContext, SFeatureAccess>(
+            AntlrScriptParser.Feature_access_expressionContext::instance to SFeatureAccess::container
+        )
+        val expectedAST = SScript(
+            mutableListOf(
+                SCreateStatement(ReferenceByName("Client")),
+                SSetStatement(
+                    ReferenceByName("name"),
+                    SInstanceById(
+                        ReferenceByName("Client"),
+                        SIntegerLiteral(1)
+                    ),
+                    SStringLiteral("ACME Inc.")
+                ),
+                SCreateStatement(ReferenceByName("Product")),
+                SSetStatement(
+                    ReferenceByName("value"),
+                    SInstanceById(
+                        ReferenceByName("Product"),
+                        SIntegerLiteral(2)
+                    ),
+                    SDivision(SSum(SIntegerLiteral(1500), SIntegerLiteral(200)), SIntegerLiteral(2))
+                ),
+                SPrintStatement(
+                    SConcat(
+                        SStringLiteral("Value of Product #2 is: "),
+                        SFeatureAccess(
+                            ReferenceByName("value"),
+                            SInstanceById(
+                                ReferenceByName("Product"),
+                                SIntegerLiteral(2)
+                            )
+                        )
+                    )
+                )
+            )
+        )
+        val actualAST = transformer.transform(
+            parseScript(
+                """create Client
+                set name of Client #1 to 'ACME Inc.'
+                create Product
+                set value of Product #2 to (1500 + 200) / 2
+                print concat 'Value of Product #2 is: ' and value of Product #2
+                """.trimIndent()
+            )
+        )!!
+        actualAST.walk().forEach { parent ->
+            parent.children.forEach { child ->
+                assertSame(parent, child.parent)
+            }
+        }
+        assertASTsAreEqual(expectedAST, actualAST)
+    }
+
+    // Ensure that https://github.com/Strumenta/kolasu/issues/241 is fixed
+    @Test
+    fun transformChildFactory() {
+        val ctx = EntCtx("foo", listOf(EntCtxFeature("bar", EntCtxStringType())))
+        val transformer = EntTransformer()
+        val ast = transformer.transform(ctx)
+        assertASTsAreEqual(Ent("foo", listOf(EntFeature("bar", EntStringType()))), ast!!)
+        assertASTsAreEqual(
+            Ent("foo", listOf(EntFeature(null, null))),
+            transformer.transform(EntCtx("foo", listOf(null, EntCtxFeature(), null)))!!)
+    }
 }
+
+class EntTransformer(issues: MutableList<Issue> = mutableListOf()) :
+    ParseTreeToASTTransformer(issues, allowGenericNode = false) {
+    init {
+        registerNodeFactory(EntCtx::class) { ctx -> Ent(ctx.name) }
+            .withChild(Ent::features, EntCtx::features)
+        registerNodeFactory(EntCtxFeature::class) { ctx -> EntFeature(name = ctx.name) }
+            .withChild(EntFeature::type, EntCtxFeature::type,)
+        this.registerNodeFactory(EntCtxStringType::class, EntStringType::class)
+    }
+}
+
+data class EntCtx(val name: String, val features: List<EntCtxFeature?>)
+
+data class EntCtxFeature(
+    val name: String? = null,
+    var type: EntCtxType? = null
+)
+
+open class EntCtxType
+
+open class EntCtxPrimitiveType : EntCtxType()
+class EntCtxStringType : EntCtxPrimitiveType()
+
+data class Ent(override val name: String, var features: List<EntFeature> = listOf()) : Node(), Named
+
+data class EntFeature(
+    override val name: String? = null,
+    var type: EntType? = null
+) : Node(), PossiblyNamed
+
+open class EntType : Node()
+
+open class EntPrimitiveType : EntType()
+class EntStringType : EntPrimitiveType()
+
